Reduced and Refiend code. (could not test because game build just updated)
I'm using 33 as priority initially because the maximum player slot is 32 in a GTA Online session, so the priority can technically not be above 32. This way I'm removing one if statement and code is faster + code reduced.
And idk why I've set the function locally on the while loop so for performances obviously I set it outside the loop now, in the feature function.